Work on ball

+++ paddle.h
@@ -6,7 +6,7 @@ typedef struct paddle paddle;
 paddle *newPaddle (grid *g);
 void freePaddle ();
 
-void movePaddle (paddle *p, uint8_t dx);
-uint8_t getPaddleX (paddle *p);
+void movePaddle (paddle *p, float dx);
+float getPaddleX (paddle *p);
 
 void testPaddle (grid *g);
